[INLONG-2490][Manager]move some file
