+++ inlong-manager/manager-common/src/main/java/org/apache/inlong/manager/common/workflow/BaseWorkflowFormType.java
@@ -15,17 +15,17 @@
  * limitations under the License.
  */
 
-package org.apache.inlong.manager.service.workflow;
+package org.apache.inlong.manager.common.workflow;
 
 import com.fasterxml.jackson.annotation.JsonSubTypes;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
 import com.fasterxml.jackson.annotation.JsonTypeInfo.Id;
 import lombok.Data;
-import org.apache.inlong.manager.service.workflow.business.BusinessResourceWorkflowForm;
-import org.apache.inlong.manager.service.workflow.business.NewBusinessWorkflowForm;
-import org.apache.inlong.manager.service.workflow.business.UpdateBusinessWorkflowForm;
-import org.apache.inlong.manager.service.workflow.consumption.NewConsumptionWorkflowForm;
 import org.apache.inlong.manager.common.model.definition.ProcessForm;
+import org.apache.inlong.manager.common.workflow.bussiness.BusinessResourceWorkflowForm;
+import org.apache.inlong.manager.common.workflow.bussiness.NewBusinessWorkflowForm;
+import org.apache.inlong.manager.common.workflow.bussiness.UpdateBusinessWorkflowForm;
+import org.apache.inlong.manager.common.workflow.consumption.NewConsumptionWorkflowForm;
 
 /**
  * The main form of the process-submitted when the process is initiated
@@ -36,10 +36,8 @@ import org.apache.inlong.manager.common.model.definition.ProcessForm;
 @JsonSubTypes({
         @JsonSubTypes.Type(value = NewBusinessWorkflowForm.class, name = NewBusinessWorkflowForm.FORM_NAME),
         @JsonSubTypes.Type(value = NewConsumptionWorkflowForm.class, name = NewConsumptionWorkflowForm.FORM_NAME),
-        @JsonSubTypes.Type(value = BusinessResourceWorkflowForm.class,
-                name = BusinessResourceWorkflowForm.FORM_NAME),
-        @JsonSubTypes.Type(value = UpdateBusinessWorkflowForm.class,
-                name = UpdateBusinessWorkflowForm.FORM_NAME),
+        @JsonSubTypes.Type(value = BusinessResourceWorkflowForm.class, name = BusinessResourceWorkflowForm.FORM_NAME),
+        @JsonSubTypes.Type(value = UpdateBusinessWorkflowForm.class, name = UpdateBusinessWorkflowForm.FORM_NAME)
 })
 public abstract class BaseWorkflowFormType implements ProcessForm {
