+++ inlong-manager/manager-service/src/main/java/org/apache/inlong/manager/service/thirdpart/mq/TubeEventSelector.java
@@ -23,7 +23,7 @@ import org.apache.inlong.manager.common.event.EventSelector;
 import org.apache.inlong.manager.common.model.WorkflowContext;
 import org.apache.inlong.manager.common.model.definition.ProcessForm;
 import org.apache.inlong.manager.common.pojo.business.BusinessInfo;
-import org.apache.inlong.manager.service.workflow.business.BusinessResourceWorkflowForm;
+import org.apache.inlong.manager.common.workflow.bussiness.BusinessResourceWorkflowForm;
 
 @Slf4j
 public class TubeEventSelector implements EventSelector {
