+++ inlong-manager/manager-service/src/main/java/org/apache/inlong/manager/service/workflow/business/SuspendBusinessWorkflowDefinition.java
@@ -23,6 +23,7 @@ import org.apache.inlong.manager.common.model.definition.Process;
 import org.apache.inlong.manager.common.model.definition.ServiceTask;
 import org.apache.inlong.manager.common.model.definition.ServiceTaskType;
 import org.apache.inlong.manager.common.model.definition.StartEvent;
+import org.apache.inlong.manager.common.workflow.bussiness.UpdateBusinessWorkflowForm;
 import org.apache.inlong.manager.service.workflow.ProcessName;
 import org.apache.inlong.manager.service.workflow.ServiceTaskListenerFactory;
 import org.apache.inlong.manager.service.workflow.WorkflowDefinition;
