+++ inlong-manager/manager-service/src/test/java/org/apache/inlong/manager/service/thirdpart/hive/CreateHiveTableEventSelectorTest.java
@@ -19,8 +19,8 @@ package org.apache.inlong.manager.service.thirdpart.hive;
 
 import org.apache.inlong.manager.common.model.WorkflowContext;
 import org.apache.inlong.manager.common.pojo.business.BusinessInfo;
+import org.apache.inlong.manager.common.workflow.bussiness.BusinessResourceWorkflowForm;
 import org.apache.inlong.manager.service.ServiceBaseTest;
-import org.apache.inlong.manager.service.workflow.business.BusinessResourceWorkflowForm;
 import org.junit.Assert;
 import org.junit.Test;
 import org.springframework.beans.factory.annotation.Autowired;
